+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.24.0] - 2025-03-06
+
 ### Added
 
 - Add tenant filtering in Alloy config
@@ -471,7 +473,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add '--vintage' toggle
 - Add controller for Workload Management Cluster using cluster.x-k8s.io/v1beta1```
 
-[Unreleased]: https://github.com/giantswarm/logging-operator/compare/v0.23.0...HEAD
+[Unreleased]: https://github.com/giantswarm/logging-operator/compare/v0.24.0...HEAD
+[0.24.0]: https://github.com/giantswarm/logging-operator/compare/v0.23.0...v0.24.0
 [0.23.0]: https://github.com/giantswarm/logging-operator/compare/v0.22.0...v0.23.0
 [0.22.0]: https://github.com/giantswarm/logging-operator/compare/v0.21.0...v0.22.0
 [0.21.0]: https://github.com/giantswarm/logging-operator/compare/v0.20.1...v0.21.0